docs: add development notes
Document that this project doesn't follow the norms of modern JavaScript
packages, not because we're unaware, but because it doesn't make sense
for the usability we're targeting. We're still in the realm of ES5, so
let's package like it.

Document that specific versions and configurations of Prettier and
ESLint are expected, and that these can be made available with a Nix
shell.